requirements.txt
scikit-learn
pandas
numpy
fastapi
uvicorn
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
